feat(RegionFilter): moved files and better organized content
