+++ src/App.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import Header from './componets/Header';
 import Body from './componets/Body/Body';
-import RegionFilter from './componets/Body/filters/RegionFilter';
+import RegionFilter from './componets/Body/filters/RegionMap/RegionFilter';
 
 const App: React.FC = () => {
   return (
